requirements.txt
aiohttp
curl_cffi
Pillow